fix(backend): 🚚restructure app (#15)
* refactor: ♻️ single app for backend and frontend

This massively reduces headache with hono rpc and dockerization


* chore: 🙈 ignore blame for last commit

because i touched basically every file lol


* fix: 🩹 wrong packageManager


* refactor: 🎨 formatting


* build: ⬆️ upgrade turbo


* build(frontend): ⬆️ upgrade frontend dependencies


* fix: 🐛 commented postgres service


* docs: 📝 update docs


* build(db): ⬆️ uprade dependencies


* refactor(backend): 🔧 add basepath


* refactor(backend): 🎨 use factory functions for routers

this makes it easier to work with custom context types and exports one interface to create router instances


* chore: 🔧 biome format shadcn components

this doesn't lint them but formats the code in our style


* chore: 🔨 shortcut scripts

this makes calling package specific commands in workspace root possible


* refactor(backend): 🎨 use package.json imports field

we need this in place of a typescript path so type inferrence still works


* test(backend/test): ✅ add websocket endpoint test


* style: 🎨 formatting


* build(frontend): ⬆️ upgrade @tanstack/router-devtools


* refactor(backend): 🎨 simplify


* chore: 🙈 ignore custom copilot instructions


* refactor(backend): 🚚 rename module

This makes the inherent factory pattern clearer


* refactor(frontend): 🎨 use location origin

This makes  sure the api calls will always work on same origin


* fix(frontend): 🩹 fix url


* docs: 📝 update readme


* build: ⬆️ upgrade bun version


* build: ⬆️ upgrade dependencies


* chore: 📌 lockfiles


* refactor(frontend/hooks): 🚚 move  hooks folder

This makes sure any hooks added from shadcn and our  custom hooks are in the same  directory


* fix(frontend): 🐛 imports


* docs: 📝 update readme


---------
