+++ .gitignore
@@ -1,6 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # IDEs
+.copilot
 .idea
 .vscode/*
 !.vscode/extensions.json
@@ -28,7 +29,6 @@ build
 dist
 dist-ssr
 dev-dist
-
 
 # code gen
 routeTree.gen.ts
