+++ README.md
@@ -16,16 +16,16 @@ You will also need [Node.js](https://nodejs.org/en/download).
 ## 🏗️ Project structure
 
 We use 🚀 [Turborepo](https://turbo.build) for managing our apps/packages.
+Each package/app is 100% [TypeScript](https://www.typescriptlang.org/).
 
 ### 🌐 Apps
 
-Each app is 100% [TypeScript](https://www.typescriptlang.org/).
-
-- `/backend`: a 🔥 [Hono](https://hono.dev) app
-- `/frontend`: a ⚡ [Vite](https://vite.dev) app using
-  - ⚛️ [React](https://react.dev) for reactivity
-  - 🏝️ [Tanstack Router](https://tanstack.com/router/) for managing SPA routes
-  - 🏝️ [Tanstack react-query](https://tanstack.com/query) for managing server state on the client
+- `/app`: a monorepo-style app
+  - `/backend`: a 🔥 [Hono](https://hono.dev) server providing api routes and serving the frontend
+  - `/frontend`: a ⚡ [Vite](https://vite.dev) SPA using
+    - ⚛️ [React](https://react.dev) for reactivity
+    - 🏝️ [Tanstack Router](https://tanstack.com/router/) for managing SPA routes
+    - 🏝️ [Tanstack react-query](https://tanstack.com/query) for managing server state on the client
 
 ### 📦 Packages
 
@@ -34,7 +34,6 @@ Each app is 100% [TypeScript](https://www.typescriptlang.org/).
   - following the 🔐 [lucia-auth](https://lucia-auth.com/) tutorial
 - `/database`: package for handling db schema and access
   - using 💧 [Drizzle ORM](https://orm.drizzle.team/) with [PostgreSQL](https://www.postgresql.org/) on the server
-  - Local message Storage: TBD
 
 ## 🧰 Utilities
 
@@ -93,3 +92,15 @@ bun dev
 ```
 
 Both Tanstack libraries will render development tools to debug their features.
+
+To use Drizzle-kit CLI [commands](https://orm.drizzle.team/docs/kit-overview) run:
+
+```zsh
+bun db <COMMAND>
+```
+
+To use shadcn/ui CLI [commands](https://ui.shadcn.com/docs/cli) run:
+
+```zsh
+bun ui <COMMAND>
+```
