+++ app/Dockerfile
@@ -1,4 +1,4 @@
-ARG BUN_VERSION=1.1.34
+ARG BUN_VERSION=1.1.36
 
 ARG YARN_PKG_MANAGER="this.packageManager=\"yarn@1.22.22\""
 ARG BUN_PKG_MANAGER="this.packageManager=\"bun@${BUN_VERSION}\""
@@ -15,7 +15,7 @@ FROM bun as pruned
 ARG YARN_PKG_MANAGER
 COPY . .
 RUN bunx json -I -f package.json -e ${YARN_PKG_MANAGER}
-RUN bunx turbo prune --scope="@application-project-ws24/backend" --docker
+RUN bunx turbo prune --scope="@application-project-ws24/app" --docker
 
 # =============== INSTALL & BUILD =================
 
@@ -25,7 +25,7 @@ COPY --from=pruned /app/out/full/ .
 RUN bunx json -I -f package.json -e ${BUN_PKG_MANAGER}
 RUN bun install --production
 RUN bunx clean-modules -y "**/*.ts" "**/@types/**"
-RUN bunx turbo build --filter="@application-project-ws24/backend"
+RUN bunx turbo build --filter="@application-project-ws24/app"
 
 # ================== RELEASE ======================
 
@@ -33,5 +33,5 @@ FROM bun-alpine AS release
 USER bun
 COPY --from=builder /app .
 EXPOSE 3000/tcp
-WORKDIR /app/apps/backend
+WORKDIR /app/app
 ENTRYPOINT [ "bun", "dist/index.js" ]
