+++ app/backend/client.ts
@@ -1,0 +1,3 @@
+import type { api } from "./app";
+
+export type ApiType = typeof api;
