+++ app/backend/routes/chat/index.test.ts
@@ -1,0 +1,12 @@
+import { describe, expect, test } from "bun:test";
+import { testClient } from "hono/testing";
+import { chat } from ".";
+
+const client = testClient(chat);
+
+describe("/api/chat", () => {
+	test("GET /", async () => {
+		const res = client.index.$ws();
+		expect(res).toBeInstanceOf(WebSocket);
+	});
+});
