+++ app/frontend/components.json
