+++ app/frontend/package.json
@@ -11,14 +11,13 @@
     "ui": "bunx shadcn@latest"
   },
   "dependencies": {
-    "@application-project-ws24/backend": "workspace:*",
     "@radix-ui/react-slot": "^1.1.0",
-    "@tanstack/react-query": "^5.59.20",
-    "@tanstack/react-query-devtools": "^5.59.20",
+    "@tanstack/react-query": "^5.61.0",
+    "@tanstack/react-query-devtools": "^5.61.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
-    "hono": "4.6.10",
-    "lucide-react": "^0.456.0",
+    "hono": "4.6.11",
+    "lucide-react": "^0.460.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "dexie": "^4.0.10",
@@ -27,8 +26,8 @@
   },
   "devDependencies": {
     "@application-project-ws24/tsconfig": "workspace:*",
-    "@tanstack/router-devtools": "^1.81.4",
-    "@tanstack/router-plugin": "^1.79.0",
+    "@tanstack/router-devtools": "^1.82.1",
+    "@tanstack/router-plugin": "^1.81.9",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@vite-pwa/assets-generator": "^0.2.6",
@@ -37,7 +36,7 @@
     "globals": "^15.12.0",
     "postcss": "^8.4.49",
     "rollup-plugin-visualizer": "^5.12.0",
-    "tailwindcss": "^3.4.14",
+    "tailwindcss": "^3.4.15",
     "vite": "^5.4.11",
     "vite-plugin-pwa": "^0.21.0"
   }
