+++ app/frontend/public/pwa-192x192.png
