+++ app/frontend/public/pwa-512x512.png
