+++ app/frontend/public/pwa-64x64.png
