+++ app/frontend/src/components/message-input.tsx
