+++ app/frontend/src/lib/api.ts
@@ -1,0 +1,6 @@
+import type { ApiType } from "@server/client";
+import { hc } from "hono/client";
+
+const { api } = hc<ApiType>("http://127.0.0.1:3000");
+
+export default api;
