+++ app/frontend/src/lib/message-service.ts
