+++ app/frontend/src/routes/index.lazy.tsx
