+++ app/frontend/src/routes/test-page.lazy.tsx
@@ -1,9 +1,9 @@
+import MessageInput from "@/components/message-input";
+import { useMessageService } from "@/hooks/use-message-service";
+import type { Message } from "@/lib/message-service";
+import { messageDb } from "@/lib/message-service";
 import { createLazyFileRoute } from "@tanstack/react-router";
 import { useEffect, useState } from "react";
-import MessageInput from "../components/message-input";
-import { useMessageService } from "../lib/hooks/use-message-service";
-import type { Message } from "../lib/message-service";
-import { messageDb } from "../lib/message-service";
 
 export const Route = createLazyFileRoute("/test-page")({
 	component: TestPage,
