+++ app/package.json
@@ -1,0 +1,19 @@
+{
+	"name": "@application-project-ws24/app",
+	"module": "src/index.js",
+	"imports": {
+		"#lib/*": "./backend/lib/*.ts"
+	},
+	"scripts": {
+		"dev": "NODE_ENV=development bun run --hot backend/index.ts",
+		"build": "bun build backend/index.ts --outdir dist --target bun --minify",
+		"preview": "NODE_ENV=production bun run dist/index.js"
+	},
+	"dependencies": {
+		"hono": "4.6.11"
+	},
+	"devDependencies": {
+		"@application-project-ws24/tsconfig": "workspace:*",
+		"@types/bun": "^1.1.13"
+	}
+}
