+++ app/tsconfig.json
@@ -3,9 +3,6 @@
 	"compilerOptions": {
 		// Hono JSX
 		"jsx": "react-jsx",
-		"jsxImportSource": "hono/jsx",
-
-		// easy imports
-		"baseUrl": "./"
+		"jsxImportSource": "hono/jsx"
 	}
 }
