--- apps/backend/package.json
@@ -1,19 +1,0 @@
-{
-	"name": "@application-project-ws24/backend",
-	"module": "src/index.js",
-	"scripts": {
-		"dev": "NODE_ENV=development bun run --hot src/index.ts",
-		"build": "bun build src/index.ts --outdir dist --target bun --minify",
-		"preview": "NODE_ENV=production bun run dist/index.js"
-	},
-	"exports": {
-		"./client": "./src/client.ts"
-	},
-	"dependencies": {
-		"hono": "4.6.10"
-	},
-	"devDependencies": {
-		"@application-project-ws24/tsconfig": "workspace:*",
-		"@types/bun": "^1.1.13"
-	}
-}
