+++ package.json
@@ -5,6 +5,8 @@
     "check": "turbo biome",
     "check:fix": "turbo biome:fix",
     "build": "turbo build",
+    "db": "bun run --filter=@application-project-ws24/database db",
+    "ui": "bun run --filter=@application-project-ws24/frontend ui",
     "dev": "turbo dev",
     "preview": "turbo preview",
     "biome": "biome check",
@@ -12,16 +14,16 @@
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
-    "turbo": "^2.2.3",
+    "turbo": "^2.3.0",
     "typescript": "5.6.3"
   },
   "engines": {
-    "node": ">=20",
-    "bun": ">=1.1.33"
+    "bun": ">=1.1.36"
   },
-  "packageManager": "bun@1.1.33",
+  "packageManager": "bun@1.1.36",
   "workspaces": [
-    "apps/*",
+    "app",
+    "app/frontend",
     "packages/*"
   ],
   "trustedDependencies": [
