+++ packages/database/package.json
@@ -10,12 +10,12 @@
   },
   "devDependencies": {
     "@application-project-ws24/tsconfig": "workspace:*",
-    "@types/bun": "latest",
+    "@types/bun": "^1.1.13",
     "@types/pg": "^8.11.10",
-    "drizzle-kit": "^0.28.0"
+    "drizzle-kit": "^0.28.1"
   },
   "dependencies": {
-    "drizzle-orm": "^0.36.1",
+    "drizzle-orm": "^0.36.3",
     "pg": "^8.13.1"
   }
 }
