+++ turbo.json
@@ -5,7 +5,7 @@
 		"build": {
 			"dependsOn": ["^build"],
 			"inputs": ["$TURBO_DEFAULT$", ".env*"],
-			"outputs": ["dist"]
+			"outputs": ["dist", "../dist/client"]
 		},
 		"dev": {
 			"cache": false,
